add ownership to assignment

+++ src/main/java/tech/stark/webapp/service/AssignmentService.java
@@ -1,14 +1,17 @@
 package tech.stark.webapp.service;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Component;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.stereotype.Service;
 import tech.stark.webapp.models.Assignment;
 import tech.stark.webapp.repository.AssignmentRepository;
+import tech.stark.webapp.security.SecurityService;
 
+import java.time.Instant;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
@@ -17,6 +20,12 @@ import java.util.stream.StreamSupport;
 @Service
 public class AssignmentService {
 
+    private final Logger LOGGER = LoggerFactory.getLogger(AssignmentService.class);
+    @Autowired
+    private SecurityService securityService;
+
+    @Autowired
+    private AccountService accountService;
     @Autowired
     private AssignmentRepository assignmentRepository;
 
@@ -33,22 +42,40 @@ public class AssignmentService {
     }
 
     public Assignment save(Assignment assignment) {
+        assignment.setUser_email(securityService.getUser().getUsername());
+        assignment.setAssignment_created(Instant.now().toString());
+        assignment.setAssignment_updated(Instant.now().toString());
         return assignmentRepository.save(assignment);
     }
 
-    public boolean deleteAssigment(String id) {
+    public Optional<Boolean> deleteAssigment(String id) {
         if(assignmentRepository.existsById(id)){
-            assignmentRepository.deleteById(id);
-            return true;
+            Assignment oldAssignment = assignmentRepository.findById(id).get();
+            if(oldAssignment.getUser_email().equals(securityService.getUser().getUsername())){
+                assignmentRepository.deleteById(id);
+                return Optional.of(true);
+            } else {
+                return null;
+            }
+
         } else {
-            return false;
+            return Optional.of(false);
         }
     }
 
     public Optional<Assignment> updateAssignment(String id, Assignment assignment) {
         if(assignmentRepository.existsById(id)){
             assignment.setId(id);
-            assignmentRepository.save(assignment);
+            Assignment oldAssignment = assignmentRepository.findById(id).get();
+            if(oldAssignment.getUser_email().equals(securityService.getUser().getUsername())){
+                assignment.setAssignment_created(oldAssignment.getAssignment_created());
+                assignment.setUser_email(oldAssignment.getUser_email());
+                assignment.setAssignment_updated(Instant.now().toString());
+                assignmentRepository.save(assignment);
+            } else {
+                return null;
+            }
+
             return Optional.of(assignment);
         } else {
             return Optional.empty();
